feat: update metadata and add new HeroSection and ScrollToFeatures components for improved prompt management experience
