+++ src/app/layout.tsx
@@ -18,8 +18,27 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "PromptBear - Organize Your Prompts Like a Pro",
+  description: "Smart prompt management platform with team collaboration, parameterized prompts, and instant access. Organize, share, and optimize your AI prompts efficiently.",
+  keywords: ["prompt management", "AI prompts", "team collaboration", "prompt organization", "AI tools"],
+  authors: [{ name: "PromptBear Team" }],
+  creator: "PromptBear",
+  publisher: "PromptBear",
+  openGraph: {
+    title: "PromptBear - Organize Your Prompts Like a Pro",
+    description: "Smart prompt management platform with team collaboration, parameterized prompts, and instant access.",
+    type: "website",
+    locale: "en_US",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "PromptBear - Organize Your Prompts Like a Pro",
+    description: "Smart prompt management platform with team collaboration, parameterized prompts, and instant access.",
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
 };
 
 export default function RootLayout({ children }: { children: ReactNode }) {
